address PR feedback
* move sqldatabase resource and tests to their own files
* named tuple for ParentSpec
* update SQL Server name in ARM template to match the specified date
