+++ tools/c7n_azure/c7n_azure/entry.py
@@ -24,6 +24,7 @@ import c7n_azure.resources.resourcegroup
 import c7n_azure.resources.public_ip
 import c7n_azure.resources.storage
 import c7n_azure.resources.sqlserver
+import c7n_azure.resources.sqldatabase
 import c7n_azure.resources.vm
 import c7n_azure.resources.vnet
 import c7n_azure.resources.route_table
