+++ tools/c7n_azure/tests/cassettes/SqlServerTest.test_find_by_name.yaml
@@ -22,9 +22,9 @@ interactions:
         VmVdZhZAzO2dvPfee++999577733ujudTif33/8/XGZkAWz2zkrayZ4hgKrIHz9+fB8/In7xR5dZ
         uc4/evS9X/zR22I5++jRR5e7e+Odj0YfrepqlddtkTcfPfrFH2WzRbEsmrbO2qp+Xl0US2o6XTdt
         NSuyJTW/zOumqPCpvt+0WUuAP3qVZ7Nr+vt8XZbXP7HOyuK8yGdPq0VWLF9kCzSZTtu8aZtfVDZ5
-        TXB29/Y+3dvZPRjPsjabZE0+viLUqqtmvMzbj37J6KOymmatdJZnTbtu9gh+AeTvNutJM62LFb5u
-        7s4OHjw8P9+fbB9MZg+39+8fHGw/nEx2tu/v5nsHD+/ne/v3P71b5021rqf553W1XjV3gcrvb3G5
-        S2S4LGY0uLtfFNO6aqrzdvz6F5V35evm7gDuhNBy4+CoQXu9QoMo3I9+yfd/yf8Da8SLpZ0BAAA=
+        TXB29vd293Z2H45nWZtNsiYfXxFq1VUzXubtR79k9FFZTbNWOsuzpl03ewS/APJ3m/WkmdbFCl83
+        d2cHDx6en+9Ptg8ms4fb+/cPDrYfTiY72/d3872Dh/fzvf37n96t86Za19P887par5q7QOX3t7jc
+        JTJcFjMa3N0vimldNdV5O379i8q78nVzdwB3Qmi5cXDUoL1eoUEU7ke/5Pu/5P8BT94ZuZ0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -33,7 +33,7 @@ interactions:
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Mon, 22 Apr 2019 19:49:51 GMT
+      - Tue, 23 Apr 2019 16:43:55 GMT
       Expires:
       - '-1'
       Pragma:
@@ -49,13 +49,13 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       x-ms-correlation-request-id:
-      - a2861ab7-baef-4d11-9d81-48d68587cb84
+      - 18d1e6e6-592b-447a-8d1a-e04eed78bb43
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11994'
+      - '11998'
       x-ms-request-id:
-      - 3b830ff1-00eb-435c-8a11-67b4f9559e26
+      - a12ee6f8-ec5c-4be2-bd19-86b685984cde
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194951Z:a2861ab7-baef-4d11-9d81-48d68587cb84
+      - WESTUS2:20190423T164356Z:18d1e6e6-592b-447a-8d1a-e04eed78bb43
     status:
       code: 200
       message: OK
