+++ tools/c7n_azure/tests/cassettes/SqlServerTest.test_metric_database.yaml
@@ -22,9 +22,9 @@ interactions:
         VmVdZhZAzO2dvPfee++999577733ujudTif33/8/XGZkAWz2zkrayZ4hgKrIHz9+fB8/In7xR5dZ
         uc4/evS9X/zR22I5++jRR5e7e+Odj0YfrepqlddtkTcfPfrFH2WzRbEsmrbO2qp+Xl0US2o6XTdt
         NSuyJTW/zOumqPCpvt+0WUuAP3qVZ7Nr+vt8XZbXP7HOyuK8yGdPq0VWLF9kCzSZTtu8aZtfVDZ5
-        TXB29/Y+3dvZPRjPsjabZE0+viLUqqtmvMzbj37J6KOymmatdJZnTbtu9gh+AeTvNutJM62LFb5u
-        7s4OHjw8P9+fbB9MZg+39+8fHGw/nEx2tu/v5nsHD+/ne/v3P71b5021rqf553W1XjV3gcrvb3G5
-        S2S4LGY0uLtfFNO6aqrzdvz6F5V35evm7gDuhNBy4+CoQXu9QoMo3I9+yfd/yf8Da8SLpZ0BAAA=
+        TXB29vd293Z2H45nWZtNsiYfXxFq1VUzXubtR79k9FFZTbNWOsuzpl03ewS/APJ3m/WkmdbFCl83
+        d2cHDx6en+9Ptg8ms4fb+/cPDrYfTiY72/d3872Dh/fzvf37n96t86Za19P887par5q7QOX3t7jc
+        JTJcFjMa3N0vimldNdV5O379i8q78nVzdwB3Qmi5cXDUoL1eoUEU7ke/5Pu/5P8BT94ZuZ0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -33,7 +33,7 @@ interactions:
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Mon, 22 Apr 2019 19:49:51 GMT
+      - Tue, 23 Apr 2019 16:43:57 GMT
       Expires:
       - '-1'
       Pragma:
@@ -49,13 +49,13 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       x-ms-correlation-request-id:
-      - f8a67616-599a-4a82-8c7e-8d8adfab5f25
+      - e1dce4bb-8b84-40f3-ad31-b2bff9853d31
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11995'
+      - '11999'
       x-ms-request-id:
-      - c3f1817b-cea6-4fc6-ac6a-4a49d105ed0a
+      - dfbd8ef5-1de6-4709-8c80-1e598e629993
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194952Z:f8a67616-599a-4a82-8c7e-8d8adfab5f25
+      - WESTUS2:20190423T164357Z:e1dce4bb-8b84-40f3-ad31-b2bff9853d31
     status:
       code: 200
       message: OK
@@ -74,7 +74,7 @@ interactions:
       accept-language:
       - en-US
     method: GET
-    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver12262018/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=DatabaseResourceId%20eq%20%27%2A%27&api-version=2018-01-01
+    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver04212019/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=DatabaseResourceId%20eq%20%27%2A%27&api-version=2018-01-01
   response:
     body:
       string: !!binary |
@@ -83,14 +83,13 @@ interactions:
         aT96tDP6qC0WebPKlh89+mhvZ/fh9s7+9u7Bm939R7s7j3Z2fuqu+XBv13340eijYtnm9WVW0msv
         d5/SB/T7Ov/o0fd+8UfFjD6826wnzbQuVm1RLZu7s4MHD8/P9yfbB5PZw+39+wcH2w8nk53t+7v5
         3sHD+/ne/v1P79Z5U63raf55Xa1Xzd02b9rfv/lFZUMd5fXdVV1dFrO8bu5+UUzrqqnO2/HrX1Te
-        la+bu9MpXrDtd/f2PiXcD6LvnS2b4mLeNncXeVsXU8KvXf/+U0J0vWCEf/9VTngsWxpXe72iYX00
-        /C61WWYLavOLDQ0+GoZWVtOsLH6Qz35Smz5981WqX2cX+Ue/ZPTRelnQzHz0Uj6kd3iG8rrIGyYv
-        dZvNsjbjzuSjbv9P6etJ1uSvlKBnMwLT77rfiro3MO6+/urJ65NXZy/fnH354vXdpzSBz57tP9k+
-        ePLUTOCTJ5jAU0zgKU/gq9PXX3716uT081dffvXy9d03p6/f/P6vf+L569NXP3n66u7LV1/+5NnT
-        01ev735xdvLqy9dfPnszpm/vytev756c4AX6RD6wE/j0+M3xk+PXp6bF0ycf/ZLvjz4CDXj0oM/r
-        NlusCGt6gbnVZ2Eae0YMAfI+2hnv/JLRwCsPB1/Z+XT/wc7BfWr54OG9B3t7QyD2doZBEIRP9+ht
-        /ffBL/k+oPw8nswvjl+/OX31w57KoVc2TB3NFP0nUk6ackqfeurAU0H0nlFhdX5Bgk8N86xp183e
-        R7/k/wFmzJeEcwUAAA==
+        la+bu9MpXrDtd/b3doF79L2zZVNczNvm7iJv62JK+LXr339KiK4XjPDvv8oJj2VL42qvVzSsj4bf
+        pTbLbEFtfrGhwUfD0MpqmpXFD/LZT2rTp2++SvXr7CL/6JeMPlovC5qZj17Kh/QOz1BeF3nD5KVu
+        s1nWZtyZfNTt/yl9Pcma/JUS9GxGYPpd91tR9wbG3ddfPXl98urs5ZuzL1+8vvuUJvDZs/0n2wdP
+        npoJfPIEE3iKCTzlCXx1+vrLr16dnH7+6suvXr6+++b09Zvf//VPPH99+uonT1/dffnqy588e3r6
+        6vXdL85OXn35+stnb8b07V35+vXdkxO8QJ/IB3YCnx6/OX5y/PqU3jt+/eb01Ue/5Pujj0ABHjuo
+        87rNFivCGc3Bqz4D08gzYgcQ99HOeOeXjAZeefjer+ztDL/yffwnjEHCNaVPPQ7yuJbeM1xf5xfE
+        K9Qwz5p23ex99Ev+H7SX2KOmAwAA
     headers:
       Cache-Control:
       - no-cache
@@ -99,7 +98,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Mon, 22 Apr 2019 19:49:53 GMT
+      - Tue, 23 Apr 2019 16:43:57 GMT
       Server:
       - Microsoft-IIS/10.0
       Strict-Transport-Security:
@@ -115,13 +114,13 @@ interactions:
       X-Powered-By:
       - ASP.NET
       x-ms-correlation-request-id:
-      - 49a88dd6-8787-48c2-8ebc-20e4da68a3b9
+      - 4d75cff9-9fff-43d4-b5be-60b86fbb2bdc
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11988'
+      - '11998'
       x-ms-request-id:
-      - '{49a88dd6-8787-48c2-8ebc-20e4da68a3b9}'
+      - '{4d75cff9-9fff-43d4-b5be-60b86fbb2bdc}'
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194954Z:49a88dd6-8787-48c2-8ebc-20e4da68a3b9
+      - WESTUS2:20190423T164357Z:4d75cff9-9fff-43d4-b5be-60b86fbb2bdc
     status:
       code: 200
       message: OK
