+++ tools/c7n_azure/tests/cassettes/SqlServerTest.test_metric_elastic_exclude.yaml
@@ -22,9 +22,9 @@ interactions:
         VmVdZhZAzO2dvPfee++999577733ujudTif33/8/XGZkAWz2zkrayZ4hgKrIHz9+fB8/In7xR5dZ
         uc4/evS9X/zR22I5++jRR5e7e+Odj0YfrepqlddtkTcfPfrFH2WzRbEsmrbO2qp+Xl0US2o6XTdt
         NSuyJTW/zOumqPCpvt+0WUuAP3qVZ7Nr+vt8XZbXP7HOyuK8yGdPq0VWLF9kCzSZTtu8aZtfVDZ5
-        TXB29/Y+3dvZPRjPsjabZE0+viLUqqtmvMzbj37J6KOymmatdJZnTbtu9gh+AeTvNutJM62LFb5u
-        7s4OHjw8P9+fbB9MZg+39+8fHGw/nEx2tu/v5nsHD+/ne/v3P71b5021rqf553W1XjV3gcrvb3G5
-        S2S4LGY0uLtfFNO6aqrzdvz6F5V35evm7gDuhNBy4+CoQXu9QoMo3I9+yfd/yf8Da8SLpZ0BAAA=
+        TXB29vd293Z2H45nWZtNsiYfXxFq1VUzXubtR79k9FFZTbNWOsuzpl03ewS/APJ3m/WkmdbFCl83
+        d2cHDx6en+9Ptg8ms4fb+/cPDrYfTiY72/d3872Dh/fzvf37n96t86Za19P887par5q7QOX3t7jc
+        JTJcFjMa3N0vimldNdV5O379i8q78nVzdwB3Qmi5cXDUoL1eoUEU7ke/5Pu/5P8BT94ZuZ0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -33,7 +33,7 @@ interactions:
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Mon, 22 Apr 2019 19:49:54 GMT
+      - Tue, 23 Apr 2019 16:43:57 GMT
       Expires:
       - '-1'
       Pragma:
@@ -49,13 +49,13 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       x-ms-correlation-request-id:
-      - 67802c4f-45f8-4287-94e7-0f92118de89f
+      - 0f3da941-0e4d-4fa3-bb10-dae31ffbd61f
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11975'
+      - '11999'
       x-ms-request-id:
-      - 1cf192b3-13a2-4ded-9103-cc0c00a34e97
+      - 14ffbca8-5665-4c47-a0a8-9883067ea26a
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194954Z:67802c4f-45f8-4287-94e7-0f92118de89f
+      - WESTUS2:20190423T164358Z:0f3da941-0e4d-4fa3-bb10-dae31ffbd61f
     status:
       code: 200
       message: OK
@@ -74,7 +74,7 @@ interactions:
       accept-language:
       - en-US
     method: GET
-    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver12262018/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=ElasticPoolResourceId%20eq%20%27%2A%27&api-version=2018-01-01
+    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver04212019/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=ElasticPoolResourceId%20eq%20%27%2A%27&api-version=2018-01-01
   response:
     body:
       string: !!binary |
@@ -83,9 +83,9 @@ interactions:
         aT96tDP6qC0WebPKlh89+mhvZ/fh9s7+9u7Bm939R7s7j3Z2fuqu+XBv13340eijYtnm9WVW0msv
         d5/SB/T7Ov/o0fd+8UfFjD6826wnzbQuVm1RLZu7s4MHD8/P9yfbB5PZw+39+wcH2w8nk53t+7v5
         3sHD+/ne/v1P79Z5U63raf55Xa1Xzd02b9rfv/lFZUMd5fXdVV1dFrO8bu5+UUzrqqnO2/HrX1Te
-        la+bu9MpXrDtd/f2PiXcD6LvnS2b4mLeNncXeVsXU8KvXf/+U0J0vWCEf/9VTngsWxpXe72iYX00
-        /C61WWYLavOLDQ0+GoZWVtOsLH6Qz35Smz5981WqX2cX+Ue/ZPTRelnQzHz0Uj6kd3iG8rrIGyLv
-        93/J96U/mrMpADjEPGLQW4aYdX5BKFDDPGvadbP30S/5fwC5wRa8/QEAAA==
+        la+bu9MpXrDtd/b3doF79L2zZVNczNvm7iJv62JK+LXr339KiK4XjPDvv8oJj2VL42qvVzSsj4bf
+        pTbLbEFtfrGhwUfD0MpqmpXFD/LZT2rTp2++SvXr7CL/6JeMPlovC5qZj17Kh/QOz1BeF3lD5P3+
+        L/m+9EdzNgUAh5hHDHrLELPOLwgFaphnTbtu9j76Jf8PZdFRqf0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -94,7 +94,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Mon, 22 Apr 2019 19:49:54 GMT
+      - Tue, 23 Apr 2019 16:44:00 GMT
       Server:
       - Microsoft-IIS/10.0
       Strict-Transport-Security:
@@ -110,13 +110,13 @@ interactions:
       X-Powered-By:
       - ASP.NET
       x-ms-correlation-request-id:
-      - 7af2954a-4741-4d67-8f6c-fe44a1564068
+      - a77d07cf-bd48-4736-949f-189bae24a0f5
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11981'
+      - '11997'
       x-ms-request-id:
-      - '{7af2954a-4741-4d67-8f6c-fe44a1564068}'
+      - '{a77d07cf-bd48-4736-949f-189bae24a0f5}'
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194954Z:7af2954a-4741-4d67-8f6c-fe44a1564068
+      - WESTUS2:20190423T164400Z:a77d07cf-bd48-4736-949f-189bae24a0f5
     status:
       code: 200
       message: OK
