+++ tools/c7n_azure/tests/cassettes/SqlServerTest.test_metric_elastic_include.yaml
@@ -22,9 +22,9 @@ interactions:
         VmVdZhZAzO2dvPfee++999577733ujudTif33/8/XGZkAWz2zkrayZ4hgKrIHz9+fB8/In7xR5dZ
         uc4/evS9X/zR22I5++jRR5e7e+Odj0YfrepqlddtkTcfPfrFH2WzRbEsmrbO2qp+Xl0US2o6XTdt
         NSuyJTW/zOumqPCpvt+0WUuAP3qVZ7Nr+vt8XZbXP7HOyuK8yGdPq0VWLF9kCzSZTtu8aZtfVDZ5
-        TXB29/Y+3dvZPRjPsjabZE0+viLUqqtmvMzbj37J6KOymmatdJZnTbtu9gh+AeTvNutJM62LFb5u
-        7s4OHjw8P9+fbB9MZg+39+8fHGw/nEx2tu/v5nsHD+/ne/v3P71b5021rqf553W1XjV3gcrvb3G5
-        S2S4LGY0uLtfFNO6aqrzdvz6F5V35evm7gDuhNBy4+CoQXu9QoMo3I9+yfd/yf8Da8SLpZ0BAAA=
+        TXB29vd293Z2H45nWZtNsiYfXxFq1VUzXubtR79k9FFZTbNWOsuzpl03ewS/APJ3m/WkmdbFCl83
+        d2cHDx6en+9Ptg8ms4fb+/cPDrYfTiY72/d3872Dh/fzvf37n96t86Za19P887par5q7QOX3t7jc
+        JTJcFjMa3N0vimldNdV5O379i8q78nVzdwB3Qmi5cXDUoL1eoUEU7ke/5Pu/5P8BT94ZuZ0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -33,7 +33,7 @@ interactions:
       Content-Type:
       - application/json; charset=utf-8
       Date:
-      - Mon, 22 Apr 2019 19:49:55 GMT
+      - Tue, 23 Apr 2019 16:44:00 GMT
       Expires:
       - '-1'
       Pragma:
@@ -49,13 +49,13 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       x-ms-correlation-request-id:
-      - f9f736e5-cbc7-4336-9d91-5b3bd36d4c9c
+      - 55461f27-845a-405c-990c-75259713a7e6
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11984'
+      - '11999'
       x-ms-request-id:
-      - 564cb9f5-feb4-4679-8419-f1e7f9b0dbce
+      - 23e06273-8bf2-40e1-8718-11a3fe254559
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194955Z:f9f736e5-cbc7-4336-9d91-5b3bd36d4c9c
+      - WESTUS2:20190423T164400Z:55461f27-845a-405c-990c-75259713a7e6
     status:
       code: 200
       message: OK
@@ -74,7 +74,7 @@ interactions:
       accept-language:
       - en-US
     method: GET
-    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver12262018/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=ElasticPoolResourceId%20eq%20%27%2A%27&api-version=2018-01-01
+    uri: https://management.azure.com/subscriptions/ea42f556-5106-4743-99b0-c129bfa71a47/resourceGroups/test_sqlserver/providers/Microsoft.Sql/servers/cctestsqlserver04212019/providers/microsoft.insights/metrics?timespan=2019-04-18%2014%3A10%3A00%2F2019-04-21%2014%3A10%3A00&interval=P1D&metricnames=dtu_consumption_percent&aggregation=average&$filter=ElasticPoolResourceId%20eq%20%27%2A%27&api-version=2018-01-01
   response:
     body:
       string: !!binary |
@@ -83,9 +83,9 @@ interactions:
         aT96tDP6qC0WebPKlh89+mhvZ/fh9s7+9u7Bm939R7s7j3Z2fuqu+XBv13340eijYtnm9WVW0msv
         d5/SB/T7Ov/o0fd+8UfFjD6826wnzbQuVm1RLZu7s4MHD8/P9yfbB5PZw+39+wcH2w8nk53t+7v5
         3sHD+/ne/v1P79Z5U63raf55Xa1Xzd02b9rfv/lFZUMd5fXdVV1dFrO8bu5+UUzrqqnO2/HrX1Te
-        la+bu9MpXrDtd/f2PiXcD6LvnS2b4mLeNncXeVsXU8KvXf/+U0J0vWCEf/9VTngsWxpXe72iYX00
-        /C61WWYLavOLDQ0+GoZWVtOsLH6Qz35Smz5981WqX2cX+Ue/ZPTRelnQzHz0Uj6kd3iG8rrIGyLv
-        93/J96U/mrMpADjEPGLQW4aYdX5BKFDDPGvadbP30S/5fwC5wRa8/QEAAA==
+        la+bu9MpXrDtd/b3doF79L2zZVNczNvm7iJv62JK+LXr339KiK4XjPDvv8oJj2VL42qvVzSsj4bf
+        pTbLbEFtfrGhwUfD0MpqmpXFD/LZT2rTp2++SvXr7CL/6JeMPlovC5qZj17Kh/QOz1BeF3lD5P3+
+        L/m+9EdzNgUAh5hHDHrLELPOLwgFaphnTbtu9j76Jf8PZdFRqf0BAAA=
     headers:
       Cache-Control:
       - no-cache
@@ -94,7 +94,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Mon, 22 Apr 2019 19:49:55 GMT
+      - Tue, 23 Apr 2019 16:44:00 GMT
       Server:
       - Microsoft-IIS/10.0
       Strict-Transport-Security:
@@ -110,13 +110,13 @@ interactions:
       X-Powered-By:
       - ASP.NET
       x-ms-correlation-request-id:
-      - 50c341c4-7cbf-49fc-811a-81ae655f442f
+      - 3c4f9fcb-a712-482c-83db-a2ca2c92ed54
       x-ms-ratelimit-remaining-subscription-reads:
-      - '11990'
+      - '11999'
       x-ms-request-id:
-      - '{50c341c4-7cbf-49fc-811a-81ae655f442f}'
+      - '{3c4f9fcb-a712-482c-83db-a2ca2c92ed54}'
       x-ms-routing-request-id:
-      - WESTUS2:20190422T194955Z:50c341c4-7cbf-49fc-811a-81ae655f442f
+      - WESTUS2:20190423T164401Z:3c4f9fcb-a712-482c-83db-a2ca2c92ed54
     status:
       code: 200
       message: OK
